Bug Fixes and Changes
--------CitizensCommands.java--------
Added a value flag for delay times on commands (cmdadd, cmdset)
Fixed a possible issue of when you enter anything other than a number on
value flag 'v' within the cmdadd command.

--------CommandDatabase.java--------
Added a save point for the delay (Is backwards compatible with all
versions)
Changed spacer from '~' to '~~~' (As a request as this was interfering
with commands)

--------Config.java--------
Fixed bug where ExecuteCommandMessage would always be true.

--------NPCListener.java && NPCCommand.java--------
Implemented the delay factor

--------plugin.yml--------
Changed version to v1.8.7