Multiple Updates
Added random flag to command. (If one of the commands has the random
flag it will pick randomly from all the commands and only execute that
one per hit on the NPC)
Added customization of messages sent out from the plugin.
Added the ability to remove the cooldown message.
Changed the way the configuration works in the background.